requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
werkzeug>=3.0.0
pandas
openpyxl
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests
